# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Foot Easy API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/invitations": {
      "get": {
        "tags": [
          "Invitations"
        ],
        "summary": "List Invitations",
        "description": "List an event's invitations with each member's availability.",
        "operationId": "list_invitations_api_v1_invitations_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "availability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Availability"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Availability"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_InvitationResponse_"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invitations"
        ],
        "summary": "Create Invitations",
        "description": "Invite members (or the whole roster) to an event; returns only new invitations.",
        "operationId": "create_invitations_api_v1_invitations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationResponse"
                  },
                  "title": "Response Create Invitations Api V1 Invitations Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/summary": {
      "get": {
        "tags": [
          "Invitations"
        ],
        "summary": "Get Availability Summary",
        "description": "Count an event's invitations per availability.",
        "operationId": "get_availability_summary_api_v1_invitations_summary_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilitySummary"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/reminders": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "summary": "Create Reminders",
        "description": "Remind every member who has not answered the event's invitation yet.",
        "operationId": "create_reminders_api_v1_invitations_reminders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResult"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invitations/{invitation_id}": {
      "patch": {
        "tags": [
          "Invitations"
        ],
        "summary": "Partial Update Invitation",
        "description": "Record a member's availability for the event.",
        "operationId": "partial_update_invitation_api_v1_invitations__invitation_id__patch",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationReply"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Invitations"
        ],
        "summary": "Delete Invitation",
        "description": "Withdraw an invitation.",
        "operationId": "delete_invitation_api_v1_invitations__invitation_id__delete",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lineups/formations": {
      "get": {
        "tags": [
          "Lineups"
        ],
        "summary": "List Formations",
        "description": "List the supported tactical formations (11, 8 and 5-a-side).",
        "operationId": "list_formations_api_v1_lineups_formations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FormationResponse"
                  },
                  "type": "array",
                  "title": "Response List Formations Api V1 Lineups Formations Get"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lineups/{event_id}": {
      "get": {
        "tags": [
          "Lineups"
        ],
        "summary": "Get Lineup",
        "description": "Get an event's lineup, flagging selected members who are not available.",
        "operationId": "get_lineup_api_v1_lineups__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lineups"
        ],
        "summary": "Update Lineup",
        "description": "Create or fully replace an event's lineup (formation, starters, substitutes).",
        "operationId": "update_lineup_api_v1_lineups__event_id__put",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineupWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lineups"
        ],
        "summary": "Delete Lineup",
        "description": "Delete an event's lineup.",
        "operationId": "delete_lineup_api_v1_lineups__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Availability": {
        "type": "string",
        "enum": [
          "pending",
          "available",
          "uncertain",
          "unavailable"
        ],
        "title": "Availability"
      },
      "AvailabilitySummary": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "invited": {
            "type": "integer",
            "title": "Invited"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "uncertain": {
            "type": "integer",
            "title": "Uncertain"
          },
          "unavailable": {
            "type": "integer",
            "title": "Unavailable"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "invited",
          "pending",
          "available",
          "uncertain",
          "unavailable"
        ],
        "title": "AvailabilitySummary"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError"
      },
      "FormationResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "players": {
            "type": "integer",
            "title": "Players"
          },
          "lines": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "code",
          "players",
          "lines"
        ],
        "title": "FormationResponse"
      },
      "InvitationCreate": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "member_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Ids",
            "description": "Members to invite; the whole roster when omitted"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "InvitationCreate"
      },
      "InvitationReply": {
        "properties": {
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "uncertain",
              "unavailable"
            ],
            "title": "Availability"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "availability"
        ],
        "title": "InvitationReply"
      },
      "InvitationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "responded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At"
          },
          "reminder_count": {
            "type": "integer",
            "title": "Reminder Count"
          },
          "last_reminded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reminded At"
          },
          "member": {
            "$ref": "#/components/schemas/InvitedMember"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "availability",
          "comment",
          "responded_at",
          "reminder_count",
          "last_reminded_at",
          "member"
        ],
        "title": "InvitationResponse"
      },
      "InvitedMember": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "shirt_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shirt Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role",
          "position",
          "shirt_number"
        ],
        "title": "InvitedMember"
      },
      "LineupResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "formation": {
            "type": "string",
            "title": "Formation"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotResponse"
            },
            "type": "array",
            "title": "Slots"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "unavailable_member_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Unavailable Member Ids",
            "description": "Selected members who did not answer 'available'",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_id",
          "formation",
          "is_published",
          "slots",
          "updated_at"
        ],
        "title": "LineupResponse"
      },
      "LineupWrite": {
        "properties": {
          "formation": {
            "type": "string",
            "title": "Formation",
            "examples": [
              "4-4-2"
            ]
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": false
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Slots"
          }
        },
        "type": "object",
        "required": [
          "formation",
          "slots"
        ],
        "title": "LineupWrite"
      },
...
      "MemberRole": {
        "type": "string",
        "enum": [
          "player",
          "coach",
          "staff"
        ],
        "title": "MemberRole"
      },
...
      "PaginatedResponse_InvitationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvitationResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[InvitationResponse]"
      },
...
      "PlayerPosition": {
        "type": "string",
        "enum": [
          "goalkeeper",
          "defender",
          "midfielder",
          "forward"
        ],
        "title": "PlayerPosition"
      },
...
      "ReminderRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          }
        },
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "ReminderRequest"
      },
      "ReminderResult": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "reminded": {
            "type": "integer",
            "title": "Reminded"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "reminded"
        ],
        "title": "ReminderResult"
      },
...
      "SlotResponse": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/SlotRole"
          },
          "position_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Index"
          },
          "member": {
            "$ref": "#/components/schemas/InvitedMember"
          }
        },
        "type": "object",
        "required": [
          "role",
          "position_index",
          "member"
        ],
        "title": "SlotResponse"
      },
      "SlotRole": {
        "type": "string",
        "enum": [
          "starter",
          "substitute"
        ],
        "title": "SlotRole"
      },
...
    }
  }
}